\name{fortify}
\alias{fortify}
\title{Convert any object to a data frame, suitable for visualisation.}
\usage{fortify(model, data, ...)}

\description{
  Convert any object to a data frame, suitable for
  visualisation.
}
\seealso{Fortify methods: \Sexpr[results=rd]{fortify:::fortify_methods()}}
\arguments{
  \item{model}{model, or other object}
  \item{data}{optional original dataset}
  \item{...}{other arguments passed on to individual methods}
}
